Modified scaling of int24_t::asDouble(). Added sorting of velocity layers in WaveSplitter

+++ AudioPlayer/DfxAudio.cpp
@@ -1030,7 +1030,7 @@ namespace dfx
                             for (int j = 0; j < nChannels; j++)
                             {
                                 auto& d = out[info.outOffset[j]];
-                                d = in[info.inOffset[j]].asDouble();
+                                d = double(in[info.inOffset[j]].asInt());
                                 d += 0.5;
                                 d *= scale;
                             }
